requirements.txt
playwright>=1.40.0
openpyxl>=3.1.2
click>=8.1.7
rich>=13.7.0
pillow>=10.2.0
requests>=2.31.0
pyinstaller>=6.3.0
